feat: Improve ENS name and avatar resolving.
