+++ src/components/Guilds/ProposalCard/index.tsx
@@ -11,7 +11,6 @@ import 'react-loading-skeleton/dist/skeleton.css';
 import UnstyledLink from '../common/UnstyledLink';
 import { useProposal } from 'hooks/Guilds/ether-swr/useProposal';
 import useENSAvatar from '../../../hooks/Guilds/ens/useENSAvatar';
-import { useMemo } from 'react';
 import Avatar from '../Avatar';
 import { DEFAULT_ETH_CHAIN_ID } from '../../../provider/connectors';
 import { shortenAddress } from '../../../utils';
@@ -105,21 +104,10 @@ interface ProposalCardProps {
 const ProposalCard: React.FC<ProposalCardProps> = ({ id, href }) => {
   const { guild_id: guildId } = useParams<{ guild_id?: string }>();
   const { data: proposal } = useProposal(guildId, id);
-  const { avatarUri, imageUrl, ensName } = useENSAvatar(
+  const { imageUrl, ensName } = useENSAvatar(
     proposal?.creator,
     DEFAULT_ETH_CHAIN_ID
   );
-
-  const imageUrlToUse = useMemo(() => {
-    if (avatarUri) {
-      // TODO: Consider chainId when generating ENS metadata service fallback URL
-      return (
-        imageUrl || `https://metadata.ens.domains/mainnet/avatar/${ensName}`
-      );
-    } else {
-      return null;
-    }
-  }, [imageUrl, ensName, avatarUri]);
 
   return (
     <UnstyledLink to={href || '#'}>
@@ -128,7 +116,7 @@ const ProposalCard: React.FC<ProposalCardProps> = ({ id, href }) => {
           <IconDetailWrapper>
             {proposal?.creator ? (
               <Avatar
-                src={imageUrlToUse}
+                src={imageUrl}
                 defaultSeed={proposal.creator}
                 size={24}
               />
